refactor: move main components into packages
