+++ view/app/auth/[[...path]]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { ResetPasswordUsingToken } from 'supertokens-auth-react/recipe/emailpassword/prebuiltui';
-import { LoginForm } from '@/components/features/login-form';
+import { LoginForm } from '@/packages/components/login-form';
 import useAuth from '../../../packages/hooks/auth/use-auth';
 
 export default function Auth() {
