+++ view/app/containers/[id]/components/ContainerDetailsHeader.tsx
@@ -3,7 +3,7 @@
 import { Play, StopCircle, Trash2, RotateCw } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
-import { ResourceGuard } from '@/components/rbac/PermissionGuard';
+import { ResourceGuard } from '@/packages/components/rbac';
 import { Skeleton } from '@/components/ui/skeleton';
 import { cn } from '@/lib/utils';
 import SubPageHeader from '@/components/ui/sub-page-header';
