+++ view/app/self-host/create/page.tsx
@@ -1,9 +1,9 @@
 'use client';
 import React from 'react';
 import { ListRepositories } from '@/packages/components/github-repositories';
-import { ResourceGuard } from '@/components/rbac/PermissionGuard';
+import { ResourceGuard } from '@/packages/components/rbac';
 import { Skeleton } from '@/components/ui/skeleton';
-import PageLayout from '@/components/layout/page-layout';
+import PageLayout from '@/packages/layouts/page-layout';
 import MainPageHeader from '@/components/ui/main-page-header';
 
 function page() {
