+++ view/app/self-host/page.tsx
@@ -13,9 +13,9 @@ import { useTranslation } from '@/hooks/use-translation';
 import { useFeatureFlags } from '@/hooks/features_provider';
 import Skeleton from '../../packages/deprecated/file-manager/components/skeleton/Skeleton';
 import { FeatureNames } from '@/packages/types/feature-flags';
-import DisabledFeature from '@/components/features/disabled-feature';
-import { ResourceGuard, AnyPermissionGuard } from '@/components/rbac/PermissionGuard';
-import PageLayout from '@/components/layout/page-layout';
+import DisabledFeature from '@/packages/components/rbac';
+import { ResourceGuard, AnyPermissionGuard } from '@/packages/components/rbac';
+import PageLayout from '@/packages/layouts/page-layout';
 import { TypographyH2, TypographyMuted } from '@/components/ui/typography';
 import { Plus } from 'lucide-react';
 import { LabelFilter } from '@/components/ui/label-filter';
