+++ view/packages/components/general-settings.tsx
@@ -7,10 +7,10 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { useTranslation } from '@/hooks/use-translation';
-import { LanguageSwitcher } from '@/components/language-switcher';
+import { LanguageSwitcher } from '@/packages/components/language-switcher';
 import { SelectWrapper } from '@/components/ui/select-wrapper';
 import { Switch } from '@/components/ui/switch';
-import { RBACGuard } from '@/components/rbac/RBACGuard';
+import { RBACGuard } from '@/packages/components/rbac';
 import { TypographySmall, TypographyMuted, TypographyH3 } from '@/components/ui/typography';
 import UploadAvatar from '@/components/ui/upload_avatar';
 import { Badge } from '@/components/ui/badge';
