+++ view/packages/components/settings.tsx
@@ -6,7 +6,7 @@ import { LogoutDialog } from '@/components/ui/logout-dialog';
 import { Button } from '@/components/ui/button';
 import { LogOut } from 'lucide-react';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { ResourceGuard } from '@/components/rbac/PermissionGuard';
+import { ResourceGuard } from '@/packages/components/rbac';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Switch } from '@/components/ui/switch';
