+++ view/packages/components/terminal.tsx
@@ -15,9 +15,9 @@ import useTerminalPane from '../hooks/terminal/use-terminal-pane';
 import '@xterm/xterm/css/xterm.css';
 import { useTranslation } from '@/hooks/use-translation';
 import { useFeatureFlags } from '@/hooks/features_provider';
-import DisabledFeature from '@/components/features/disabled-feature';
+import DisabledFeature from '@/packages/components/rbac';
 import { FeatureNames } from '@/packages/types/feature-flags';
-import { AnyPermissionGuard } from '@/components/rbac/PermissionGuard';
+import { AnyPermissionGuard } from '@/packages/components/rbac';
 import { useRBAC } from '@/lib/rbac';
 import { ResizablePanelGroup, ResizablePanel, ResizableHandle } from '@/components/ui/resizable';
 import { useTerminalSessions } from '@/packages/hooks/terminal/use-terminal-sessions';
