+++ view/packages/deprecated/file-manager/page.tsx
@@ -18,10 +18,10 @@ import { FileData } from '@/redux/types/files';
 import { FileContextMenu } from './components/context-menu/FileContextMenu';
 import { toast } from 'sonner';
 import { useFeatureFlags } from '@/hooks/features_provider';
-import DisabledFeature from '@/components/features/disabled-feature';
+import DisabledFeature from '@/packages/components/rbac';
 import { FeatureNames } from '@/packages/types/feature-flags';
-import { ResourceGuard, AnyPermissionGuard } from '@/components/rbac/PermissionGuard';
-import PageLayout from '@/components/layout/page-layout';
+import { ResourceGuard, AnyPermissionGuard } from '@/packages/components/rbac';
+import PageLayout from '@/packages/layouts/page-layout';
 import { useIsMobile } from '@/hooks/use-mobile';
 import { DialogWrapper } from '@/components/ui/dialog-wrapper';
 import FileInfo from './components/file-list/FileInfo';
