+++ view/packages/hooks/applications/use_deployments_list.tsx
@@ -135,4 +135,3 @@ export function useDeploymentsList({ deployments }: UseDeploymentsListProps) {
     getStatusIcon
   };
 }
-
